requirements.txt
fastapi
uvicorn
requests
python-multipart
imgocr
python-dotenv
onnxruntime
